+++ type-tests/src/main.rs
@@ -24,7 +24,7 @@ use {
     thiserror::Error,
 };
 
-// const SINGLE: Option<&str> = Some("t0106");
+// const SINGLE: Option<&str> = Some("t0120");
 const SINGLE: Option<&str> = None;
 const WRITE_MISSING: bool = true;
 const WRITE_FAILED: bool = false;